refactor(code): remove the Bleve pack path and legacy code_* tools (CS6)
The seamless tools (code_map/code_outline/code_graph) fully replace the
pack-based workflow, so the pack machinery is retired:

- delete internal/code/{pack.go,registry.go,tools.go,pack_test.go} and the
  top-level cli_code.go (--pack/--packs/--refresh/--unpack CLI).
- drop those flag cases + help lines and code.RegisterTools/codeRegistry from
  main.go; keep RegisterSeamlessTools.
- trim the pack helpers from paths.go (DataDir stays).
- relocate codeTextResult/codeErrResult/lastSegment into codetools.go.

internal/code no longer imports Bleve (no full-text code index); Bleve stays
for internal/search (KB, and the coming memory store). Nothing is written into
a project repo — code context lives in ~/.detritus/code-tags. MCP surface is
now 7 tools (4 kb + 3 code); TestMCPServer updated. Full -race green.
