+++ internal/code/paths.go
@@ -4,28 +4,10 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"regexp"
 	"sync"
 )
 
 var dataDirWarnOnce sync.Once
-
-var validPackName = regexp.MustCompile(`^[A-Za-z0-9._-]+$`)
-
-// ValidatePackName rejects names that could escape the packs directory or
-// collide with special filesystem entries.
-func ValidatePackName(name string) error {
-	if name == "" {
-		return fmt.Errorf("pack name required")
-	}
-	if name == "." || name == ".." {
-		return fmt.Errorf("invalid pack name %q", name)
-	}
-	if !validPackName.MatchString(name) {
-		return fmt.Errorf("invalid pack name %q: allowed characters are [A-Za-z0-9._-]", name)
-	}
-	return nil
-}
 
 // DataDir returns the detritus data directory.
 // Resolution order: $DETRITUS_HOME, $XDG_DATA_HOME/detritus, ~/.detritus.
@@ -47,29 +29,4 @@ func DataDir() string {
 		return fallback
 	}
 	return filepath.Join(home, ".detritus")
-}
-
-// PacksDir returns the directory under which all packs live.
-func PacksDir() string {
-	return filepath.Join(DataDir(), "packs")
-}
-
-// PackDir returns the directory for a single named pack.
-func PackDir(name string) string {
-	return filepath.Join(PacksDir(), name)
-}
-
-// ManifestPath returns the path of a pack's manifest.json.
-func ManifestPath(name string) string {
-	return filepath.Join(PackDir(name), "manifest.json")
-}
-
-// IndexPath returns the directory of a pack's Bleve index.
-func IndexPath(name string) string {
-	return filepath.Join(PackDir(name), "index.bleve")
-}
-
-// EnsurePacksDir creates the packs parent dir if it doesn't exist.
-func EnsurePacksDir() error {
-	return os.MkdirAll(PacksDir(), 0o755)
 }
