+++ main_test.go
@@ -46,8 +46,8 @@ func TestMCPServer(t *testing.T) {
 	if err != nil {
 		t.Fatal("ListTools:", err)
 	}
-	if len(tools.Tools) != 12 {
-		t.Fatalf("expected 12 tools, got %d", len(tools.Tools))
+	if len(tools.Tools) != 7 {
+		t.Fatalf("expected 7 tools, got %d", len(tools.Tools))
 	}
 	toolNames := map[string]bool{}
 	for _, tool := range tools.Tools {
@@ -55,7 +55,6 @@ func TestMCPServer(t *testing.T) {
 	}
 	for _, name := range []string{
 		"kb_list", "kb_get", "kb_search", "kb_sections",
-		"code_list", "code_tree", "code_search", "code_get", "code_pack",
 		"code_map", "code_outline", "code_graph",
 	} {
 		if !toolNames[name] {
